Overerving keuken en bar
Ik heb het oude keukenscherm tot bereidingscherm gemaakt en keukenscherm
en barscherm erven dit.
